SplitConcatPairToInterpolate transformation was moved to middle stage.
