+++ model-optimizer/extensions/middle/SplitConcatPairToInterpolate_test.py
@@ -19,7 +19,7 @@ import unittest
 
 import numpy as np
 
-from extensions.front.tf.SplitConcatPairToInterpolate import SplitConcatPairToInterpolate
+from extensions.middle.SplitConcatPairToInterpolate import SplitConcatPairToInterpolate
 from mo.front.common.partial_infer.utils import int64_array
 from mo.utils.ir_engine.compare_graphs import compare_graphs
 from mo.utils.unittest.graph import build_graph
